changed product page to add type
