--- inventory-management/src/components/products/ProductForm.tsx
@@ -1,236 +1,0 @@
-import React, { useState } from 'react';
-import { X, RefreshCw } from 'lucide-react';
-import { Product } from '../../services/productService';
-import { toast } from 'react-toastify';
-
-interface ProductFormProps {
-  onSubmit: (data: any) => Promise<void>;
-  onClose: () => void;
-  initialData?: Product;
-  isSubmitting?: boolean;
-  onRefresh?: () => Promise<void>;
-}
-
-const ProductForm: React.FC<ProductFormProps> = ({
-  onSubmit,
-  onClose,
-  initialData,
-  isSubmitting = false,
-  onRefresh,
-}) => {
-  const [formData, setFormData] = useState({
-    name: initialData?.name || '',
-    description: initialData?.description || '',
-  });
-
-  const [errors, setErrors] = useState({
-    name: '',
-    description: '',
-    general: ''
-  });
-
-  const [loading, setLoading] = useState(false);
-  const [isRefreshing, setIsRefreshing] = useState(false);
-
-  const handleChange = (e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement | HTMLSelectElement>) => {
-    const { name, value } = e.target;
-    setFormData(prev => ({
-      ...prev,
-      [name]: value
-    }));
-    
-    // Clear error when field is edited
-    if (errors[name as keyof typeof errors]) {
-      setErrors(prev => ({ ...prev, [name]: '' }));
-    }
-  };
-
-  const validateForm = () => {
-    const newErrors = {
-      name: '',
-      description: '',
-      general: ''
-    };
-
-    let isValid = true;
-
-    if (!formData.name) {
-      newErrors.name = 'Product name is required';
-      isValid = false;
-    }
-
-    if (!formData.description) {
-      newErrors.description = 'Description is required';
-      isValid = false;
-    }
-
-    setErrors(newErrors);
-    return isValid;
-  };
-
-  const handleRefresh = async () => {
-    if (!onRefresh) return;
-    
-    try {
-      setIsRefreshing(true);
-      await onRefresh();
-      toast.success('Data refreshed successfully', {
-        position: "top-right",
-        autoClose: 3000,
-        hideProgressBar: false,
-        closeOnClick: true,
-        pauseOnHover: true,
-        draggable: true,
-      });
-    } catch (error) {
-      console.error('Refresh error:', error);
-      toast.error('Failed to refresh data', {
-        position: "top-right",
-        autoClose: 3000,
-        hideProgressBar: false,
-        closeOnClick: true,
-        pauseOnHover: true,
-        draggable: true,
-      });
-    } finally {
-      setIsRefreshing(false);
-    }
-  };
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    if (isSubmitting) return;
-
-    if (!validateForm()) {
-      toast.error('Please fix the errors in the form', {
-        position: "top-right",
-        autoClose: 3000,
-        hideProgressBar: false,
-        closeOnClick: true,
-        pauseOnHover: true,
-        draggable: true,
-      });
-      return;
-    }
-
-    try {
-      setLoading(true);
-      await onSubmit(formData);
-      // Success toast is now handled by the parent component
-    } catch (error) {
-      console.error('Submission error:', error);
-      const errorMessage = error instanceof Error ? error.message : 'Failed to submit form';
-      setErrors(prev => ({
-        ...prev,
-        general: errorMessage
-      }));
-      toast.error(errorMessage, {
-        position: "top-right",
-        autoClose: 3000,
-        hideProgressBar: false,
-        closeOnClick: true,
-        pauseOnHover: true,
-        draggable: true,
-      });
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  return (
-    <div className="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center z-50 overflow-y-auto p-4">
-      <div className="bg-white rounded-lg p-6 w-full max-w-md max-h-[90vh] overflow-y-auto">
-        <div className="flex justify-between items-center mb-4">
-          <h2 className="text-xl font-semibold">
-            {initialData ? 'Edit Product' : 'Add New Product'}
-          </h2>
-          <div className="flex items-center space-x-2">
-            {onRefresh && (
-              <button
-                onClick={handleRefresh}
-                className={`text-gray-500 hover:text-gray-700 ${isRefreshing ? 'animate-spin' : ''}`}
-                title="Refresh Data"
-                disabled={isRefreshing}
-              >
-                <RefreshCw size={20} />
-                Refresh
-              </button>
-            )}
-            <button
-              onClick={onClose}
-              className="text-gray-500 hover:text-gray-700"
-            >
-              <X size={24} />
-            </button>
-          </div>
-        </div>
-
-        {errors.general && (
-          <div className="mb-4 p-3 bg-red-100 text-red-700 rounded">
-            {errors.general}
-          </div>
-        )}
-
-        <form onSubmit={handleSubmit} className="space-y-4">
-          <div>
-            <label className="block text-sm font-medium text-gray-700 mb-1">
-              Name
-            </label>
-            <input
-              type="text"
-              name="name"
-              value={formData.name}
-              onChange={handleChange}
-              className="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500"
-              required
-              disabled={loading}
-              placeholder="Enter product name"
-            />
-            {errors.name && (
-              <p className="mt-1 text-sm text-red-600">{errors.name}</p>
-            )}
-          </div>
-
-          <div>
-            <label className="block text-sm font-medium text-gray-700 mb-1">
-              Description
-            </label>
-            <textarea
-              name="description"
-              value={formData.description}
-              onChange={handleChange}
-              className="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500"
-              rows={3}
-              required
-              disabled={loading}
-              placeholder="Enter product description"
-            />
-            {errors.description && (
-              <p className="mt-1 text-sm text-red-600">{errors.description}</p>
-            )}
-          </div>
-
-          <div className="flex justify-end space-x-3 pt-4">
-            <button
-              type="button"
-              onClick={onClose}
-              className="px-4 py-2 border border-gray-300 rounded-md text-gray-700 hover:bg-gray-50"
-              disabled={loading}
-            >
-              Cancel
-            </button>
-            <button
-              type="submit"
-              className="px-4 py-2 bg-blue-600 text-white rounded-md hover:bg-blue-700 disabled:opacity-50"
-              disabled={loading || isSubmitting}
-            >
-              {loading ? 'Saving...' : 'Save Product'}
-            </button>
-          </div>
-        </form>
-      </div>
-    </div>
-  );
-};
-
-export default ProductForm;
