+++ inventory-management/src/pages/products/ProductStats.tsx
@@ -1,0 +1,119 @@
+import React from "react";
+import { Package, Check, AlertTriangle, RefreshCw } from "lucide-react";
+// import { Product } from "../../services/productService";
+
+interface ProductStatsProps {
+  loading: boolean;
+  totalProducts: number;
+  activeProducts: number;
+  inactiveProducts: number;
+  lastUpdated: Date | null;
+}
+
+const ProductStats: React.FC<ProductStatsProps> = ({
+  loading,
+  totalProducts,
+  activeProducts,
+  inactiveProducts,
+  lastUpdated,
+}) => {
+  return (
+    <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-4 gap-4 md:gap-6 mb-6 md:mb-8">
+      <div className="bg-white rounded-xl shadow-sm border border-gray-200 p-4 md:p-6 transition-all hover:shadow-md">
+        <div className="flex items-center justify-between">
+          <div>
+            <p className="text-xs md:text-sm font-medium text-gray-500">
+              Total Products
+            </p>
+            <p className="text-xl md:text-2xl font-bold text-gray-800">
+              {loading ? (
+                <span className="animate-pulse">...</span>
+              ) : (
+                totalProducts
+              )}
+            </p>
+          </div>
+          <div className="w-10 h-10 md:w-12 md:h-12 bg-blue-100 rounded-full flex items-center justify-center">
+            <Package className="w-5 h-5 md:w-6 md:h-6 text-blue-600" />
+          </div>
+        </div>
+        <div className="mt-2 text-xs text-gray-500">
+          {loading ? "..." : `${activeProducts} active`}
+        </div>
+      </div>
+      
+      <div className="bg-white rounded-xl shadow-sm border border-gray-200 p-4 md:p-6 transition-all hover:shadow-md">
+        <div className="flex items-center justify-between">
+          <div>
+            <p className="text-xs md:text-sm font-medium text-gray-500">
+              Active Products
+            </p>
+            <p className="text-xl md:text-2xl font-bold text-gray-800">
+              {loading ? (
+                <span className="animate-pulse">...</span>
+              ) : (
+                activeProducts
+              )}
+            </p>
+          </div>
+          <div className="w-10 h-10 md:w-12 md:h-12 bg-green-100 rounded-full flex items-center justify-center">
+            <Check className="w-5 h-5 md:w-6 md:h-6 text-green-600" />
+          </div>
+        </div>
+        <div className="mt-2 text-xs text-gray-500">
+          {loading ? "..." : `${((activeProducts / totalProducts) * 100 || 0).toFixed(1)}% of total`}
+        </div>
+      </div>
+      
+      <div className="bg-white rounded-xl shadow-sm border border-gray-200 p-4 md:p-6 transition-all hover:shadow-md">
+        <div className="flex items-center justify-between">
+          <div>
+            <p className="text-xs md:text-sm font-medium text-gray-500">
+              Inactive Products
+            </p>
+            <p className="text-xl md:text-2xl font-bold text-gray-800">
+              {loading ? (
+                <span className="animate-pulse">...</span>
+              ) : (
+                inactiveProducts
+              )}
+            </p>
+          </div>
+          <div className="w-10 h-10 md:w-12 md:h-12 bg-amber-100 rounded-full flex items-center justify-center">
+            <AlertTriangle className="w-5 h-5 md:w-6 md:h-6 text-amber-600" />
+          </div>
+        </div>
+        <div className="mt-2 text-xs text-gray-500">
+          {loading ? "..." : `${((inactiveProducts / totalProducts) * 100 || 0).toFixed(1)}% of total`}
+        </div>
+      </div>
+      
+      <div className="bg-white rounded-xl shadow-sm border border-gray-200 p-4 md:p-6 transition-all hover:shadow-md">
+        <div className="flex items-center justify-between">
+          <div>
+            <p className="text-xs md:text-sm font-medium text-gray-500">
+              Last Updated
+            </p>
+            <p className="text-xl md:text-2xl font-bold text-gray-800">
+              {loading ? (
+                <span className="animate-pulse">...</span>
+              ) : lastUpdated ? (
+                lastUpdated.toLocaleDateString()
+              ) : (
+                "N/A"
+              )}
+            </p>
+          </div>
+          <div className="w-10 h-10 md:w-12 md:h-12 bg-purple-100 rounded-full flex items-center justify-center">
+            <RefreshCw className="w-5 h-5 md:w-6 md:h-6 text-purple-600" />
+          </div>
+        </div>
+        <div className="mt-2 text-xs text-gray-500">
+          {lastUpdated && !loading ? lastUpdated.toLocaleTimeString() : ""}
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default ProductStats;
