+++ inventory-management/src/pages/products/ProductTable.tsx
@@ -1,0 +1,362 @@
+import React from "react";
+import {
+  Calendar,
+  Eye,
+  Edit2,
+  Trash2,
+  AlertCircle,
+  ArrowLeft,
+  ArrowRight,
+} from "lucide-react";
+import { Product } from "../../services/productService";
+
+interface ProductTableProps {
+  loading: boolean;
+  error: string | null;
+  products: Product[];
+  searchTerm: string;
+  sortConfig: {
+    key: keyof Product;
+    direction: "ascending" | "descending";
+  } | null;
+  currentPage: number;
+  pageSize: number;
+  totalPages: number;
+  onRequestSort: (key: keyof Product) => void;
+  onPageChange: (newPage: number) => void;
+  onViewClick: (product: Product) => void;
+  onEditClick: (product: Product) => void;
+  onDeleteClick: (productId: number) => void;
+}
+
+const ProductTable: React.FC<ProductTableProps> = ({
+  loading,
+  error,
+  products,
+  searchTerm,
+  sortConfig,
+  currentPage,
+  pageSize,
+  totalPages,
+  onRequestSort,
+  onPageChange,
+  onViewClick,
+  onEditClick,
+  onDeleteClick,
+}) => {
+//   const getStatusIcon = (deletedAt: string | null) => {
+//     return deletedAt ? 
+//       <X className="w-4 h-4 text-red-500" /> : 
+//       <Check className="w-4 h-4 text-green-500" />;
+//   };
+
+  const getTypeText = (type: string) => {
+    return type === 'raw_material' ? 'Raw Material' : 'Finished Product';
+  };
+
+  const renderSkeleton = () => {
+    return Array(5).fill(0).map((_, i) => (
+      <tr key={`skeleton-${i}`} className="animate-pulse">
+        <td className="px-6 py-4">
+          <div className="h-4 bg-gray-200 rounded w-3/4 mb-2"></div>
+          <div className="h-3 bg-gray-200 rounded w-1/2"></div>
+        </td>
+        <td className="px-6 py-4">
+          <div className="h-4 bg-gray-200 rounded w-2/3"></div>
+        </td>
+        <td className="px-6 py-4">
+          <div className="h-4 bg-gray-200 rounded w-3/4"></div>
+        </td>
+        <td className="px-6 py-4">
+          <div className="h-4 bg-gray-200 rounded w-1/2"></div>
+        </td>
+        <td className="px-6 py-4">
+          <div className="h-4 bg-gray-200 rounded w-2/3"></div>
+        </td>
+        <td className="px-6 py-4 text-right">
+          <div className="h-6 bg-gray-200 rounded w-16 inline-block"></div>
+        </td>
+      </tr>
+    ));
+  };
+
+  return (
+    <div 
+      id="products-table-container" 
+      className="bg-white rounded-xl shadow-sm border border-gray-200 overflow-hidden"
+    >
+      <div className="overflow-x-auto">
+        <table className="min-w-full divide-y divide-gray-200">
+          <thead className="bg-gray-50">
+            <tr>
+              <th
+                className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider cursor-pointer hover:bg-gray-100"
+                onClick={() => onRequestSort("id")}
+              >
+                <div className="flex items-center">
+                  ID
+                  {sortConfig?.key === "id" && (
+                    <span className="ml-1">
+                      {sortConfig.direction === "ascending" ? "↑" : "↓"}
+                    </span>
+                  )}
+                </div>
+              </th>
+              <th
+                className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider cursor-pointer hover:bg-gray-100"
+                onClick={() => onRequestSort("name")}
+              >
+                <div className="flex items-center">
+                  Name
+                  {sortConfig?.key === "name" && (
+                    <span className="ml-1">
+                      {sortConfig.direction === "ascending" ? "↑" : "↓"}
+                    </span>
+                  )}
+                </div>
+              </th>
+              <th
+                className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider cursor-pointer hover:bg-gray-100"
+                onClick={() => onRequestSort("type")}
+              >
+                <div className="flex items-center">
+                  Type
+                  {sortConfig?.key === "type" && (
+                    <span className="ml-1">
+                      {sortConfig.direction === "ascending" ? "↑" : "↓"}
+                    </span>
+                  )}
+                </div>
+              </th>
+              <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
+                Description
+              </th>
+              <th
+                className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider cursor-pointer hover:bg-gray-100"
+                onClick={() => onRequestSort("createdAt")}
+              >
+                <div className="flex items-center">
+                  Created
+                  {sortConfig?.key === "createdAt" && (
+                    <span className="ml-1">
+                      {sortConfig.direction === "ascending" ? "↑" : "↓"}
+                    </span>
+                  )}
+                </div>
+              </th>
+              <th className="px-6 py-3 text-right text-xs font-medium text-gray-500 uppercase tracking-wider">
+                Actions
+              </th>
+            </tr>
+          </thead>
+          <tbody className="bg-white divide-y divide-gray-200">
+            {loading ? (
+              renderSkeleton()
+            ) : error ? (
+              <tr>
+                <td
+                  colSpan={6}
+                  className="px-6 py-4 text-center"
+                >
+                  <div className="flex items-center justify-center text-red-600">
+                    <AlertCircle className="w-5 h-5 mr-2" />
+                    {error}
+                  </div>
+                </td>
+              </tr>
+            ) : products.length === 0 ? (
+              <tr>
+                <td
+                  colSpan={6}
+                  className="px-6 py-4 text-center text-gray-500"
+                >
+                  No products found.{" "}
+                  {searchTerm && "Try a different search term."}
+                </td>
+              </tr>
+            ) : (
+              products.map((product) => (
+                <tr key={product.id} className="hover:bg-gray-50 transition-colors">
+                  <td className="px-6 py-4 whitespace-nowrap">
+                    <div className="text-sm font-medium text-gray-900">
+                      {product.id}
+                    </div>
+                  </td>
+                  <td className="px-6 py-4 whitespace-nowrap">
+                    <div className="text-sm font-medium text-gray-900">
+                      {product.name}
+                    </div>
+                  </td>
+                  <td className="px-6 py-4 whitespace-nowrap">
+                    <div className="text-sm text-gray-900">
+                      {getTypeText(product.type)}
+                    </div>
+                  </td>
+                  <td className="px-6 py-4">
+                    <div className="text-sm text-gray-900 line-clamp-2">
+                      {product.description || "N/A"}
+                    </div>
+                  </td>
+                  <td className="px-6 py-4 whitespace-nowrap">
+                    <div className="flex items-center">
+                      <Calendar className="flex-shrink-0 h-4 w-4 text-gray-400 mr-1" />
+                      <div className="text-sm text-gray-900">
+                        {new Date(product.createdAt).toLocaleDateString()}
+                      </div>
+                    </div>
+                  </td>
+                  <td className="px-6 py-4 whitespace-nowrap text-right text-sm font-medium">
+                    <div className="flex items-center justify-end space-x-2">
+                      <button
+                        onClick={() => onViewClick(product)}
+                        className="text-green-600 hover:text-green-900 p-1 rounded-full hover:bg-green-50"
+                        title="View Details"
+                      >
+                        <Eye size={18} />
+                      </button>
+                      
+                      <button
+                        onClick={() => onEditClick(product)}
+                        className="text-blue-600 hover:text-blue-900 p-1 rounded-full hover:bg-blue-50"
+                        title="Edit Product"
+                      >
+                        <Edit2 size={18} />
+                      </button>
+                      
+                      <button
+                        onClick={() => onDeleteClick(product.id)}
+                        className="text-red-600 hover:text-red-900 p-1 rounded-full hover:bg-red-50"
+                        title="Delete Product"
+                      >
+                        <Trash2 size={18} />
+                      </button>
+                    </div>
+                  </td>
+                </tr>
+              ))
+            )}
+          </tbody>
+        </table>
+      </div>
+
+      {products.length > 0 && (
+        <div className="bg-gray-50 px-4 py-3 flex items-center justify-between border-t border-gray-200 sm:px-6">
+          <div className="flex-1 flex justify-between sm:hidden">
+            <button
+              onClick={() => onPageChange(currentPage - 1)}
+              disabled={currentPage === 1}
+              className={`relative inline-flex items-center px-4 py-2 border border-gray-300 text-sm font-medium rounded-md ${
+                currentPage === 1
+                  ? "bg-gray-100 text-gray-400 cursor-not-allowed"
+                  : "bg-white text-gray-700 hover:bg-gray-50"
+              }`}
+            >
+              <ArrowLeft size={16} className="mr-1" />
+              Previous
+            </button>
+            <span className="text-sm text-gray-700">
+              Page {currentPage} of {totalPages}
+            </span>
+            <button
+              onClick={() => onPageChange(currentPage + 1)}
+              disabled={currentPage >= totalPages}
+              className={`relative inline-flex items-center px-4 py-2 border border-gray-300 text-sm font-medium rounded-md ${
+                currentPage >= totalPages
+                  ? "bg-gray-100 text-gray-400 cursor-not-allowed"
+                  : "bg-white text-gray-700 hover:bg-gray-50"
+              }`}
+            >
+              Next
+              <ArrowRight size={16} className="ml-1" />
+            </button>
+          </div>
+          <div className="hidden sm:flex-1 sm:flex sm:items-center sm:justify-between">
+            <div>
+              <p className="text-sm text-gray-700">
+                Showing{" "}
+                <span className="font-medium">
+                  {(currentPage - 1) * pageSize + 1}
+                </span>{" "}
+                to{" "}
+                <span className="font-medium">
+                  {Math.min(
+                    currentPage * pageSize,
+                    products.length
+                  )}
+                </span>{" "}
+                of{" "}
+                <span className="font-medium">
+                  {products.length}
+                </span>{" "}
+                results
+              </p>
+            </div>
+            <div>
+              <nav
+                className="relative z-0 inline-flex rounded-md shadow-sm -space-x-px"
+                aria-label="Pagination"
+              >
+                <button
+                  onClick={() => onPageChange(currentPage - 1)}
+                  disabled={currentPage === 1}
+                  className={`relative inline-flex items-center px-2 py-2 rounded-l-md border border-gray-300 bg-white text-sm font-medium ${
+                    currentPage === 1
+                      ? "text-gray-300 cursor-not-allowed"
+                      : "text-gray-500 hover:bg-gray-50"
+                  }`}
+                >
+                  <span className="sr-only">Previous</span>
+                  <ArrowLeft className="h-5 w-5" aria-hidden="true" />
+                </button>
+                
+                {Array.from(
+                  { length: Math.min(5, totalPages) },
+                  (_, i) => {
+                    let pageNum;
+                    if (totalPages <= 5) {
+                      pageNum = i + 1;
+                    } else if (currentPage <= 3) {
+                      pageNum = i + 1;
+                    } else if (currentPage >= totalPages - 2) {
+                      pageNum = totalPages - 4 + i;
+                    } else {
+                      pageNum = currentPage - 2 + i;
+                    }
+                    return (
+                      <button
+                        key={pageNum}
+                        onClick={() => onPageChange(pageNum)}
+                        className={`relative inline-flex items-center px-4 py-2 border text-sm font-medium ${
+                          currentPage === pageNum
+                            ? "z-10 bg-blue-50 border-blue-500 text-blue-600"
+                            : "bg-white border-gray-300 text-gray-500 hover:bg-gray-50"
+                        }`}
+                      >
+                        {pageNum}
+                      </button>
+                    );
+                  }
+                )}
+                
+                <button
+                  onClick={() => onPageChange(currentPage + 1)}
+                  disabled={currentPage >= totalPages}
+                  className={`relative inline-flex items-center px-2 py-2 rounded-r-md border border-gray-300 bg-white text-sm font-medium ${
+                    currentPage >= totalPages
+                      ? "text-gray-300 cursor-not-allowed"
+                      : "text-gray-500 hover:bg-gray-50"
+                  }`}
+                >
+                  <span className="sr-only">Next</span>
+                  <ArrowRight className="h-5 w-5" aria-hidden="true" />
+                </button>
+              </nav>
+            </div>
+          </div>
+        </div>
+      )}
+    </div>
+  );
+};
+
+export default ProductTable;
